+++ lib/appconfig.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2020
+ * (c) Copyright Ascensio System SIA 2021
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -115,6 +115,13 @@ class AppConfig {
      * @var string
      */
     private $_preview = "preview";
+
+    /**
+     * The config key for the keep versions history
+     *
+     * @var string
+     */
+    private $_versionHistory = "versionHistory";
 
     /**
      * The config key for the chat display setting
@@ -623,6 +630,26 @@ class AppConfig {
      */
     public function GetPreview() {
         return $this->config->getAppValue($this->appName, $this->_preview, "false") === "true";
+    }
+
+    /**
+     * Save keep versions history
+     *
+     * @param bool $value - version history
+     */
+    public function SetVersionHistory($value) {
+        $this->logger->info("Set keep versions history: " . json_encode($value), ["app" => $this->appName]);
+
+        $this->config->setAppValue($this->appName, $this->_versionHistory, json_encode($value));
+    }
+
+    /**
+     * Get keep versions history
+     *
+     * @return bool
+     */
+    public function GetVersionHistory() {
+        return $this->config->getAppValue($this->appName, $this->_versionHistory, "true") === "true";
     }
 
     /**